requirements.txt
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
